fix: 修复PlotPos构造函数、PlotDBStorage::getPlot异常
