+++ src/plotcraft/math/PlotPos.cc
@@ -4,6 +4,7 @@
 #include "ll/api/service/Bedrock.h"
 #include "magic_enum.hpp"
 #include "mc/enums/BlockUpdateFlag.h"
+#include "mc/math/Vec3.h"
 #include "mc/world/level/BlockPos.h"
 #include "mc/world/level/BlockSource.h"
 #include "mc/world/level/block/registry/BlockTypeRegistry.h"
@@ -12,6 +13,7 @@
 #include "plotcraft/Global.h"
 #include "plotcraft/core/TemplateManager.h"
 #include "plotcraft/data/PlotDBStorage.h"
+#include "plotcraft/math/Polygon.h"
 #include <algorithm>
 #include <cmath>
 #include <cstddef>
@@ -73,9 +75,6 @@ PlotPos::PlotPos(const Vec3& vec3) {
     mX = (int)std::floor(vec3.x / total);
     mZ = (int)std::floor(vec3.z / total);
 
-    if (data::PlotDBStorage::getInstance()._initClass(*this)) {
-        return; // 从数据库中加载数据
-    }
 
     Vec3 min, max;
     bool isValid = true;
@@ -105,6 +104,9 @@ PlotPos::PlotPos(const Vec3& vec3) {
 
 
     if (isValid) {
+        if (data::PlotDBStorage::getInstance()._initClass(*this)) {
+            return; // 从数据库中加载数据
+        }
         // 按顺时针顺序存储顶点
         mVertexs = {
             min, // 左下角
@@ -141,148 +143,19 @@ bool PlotPos::isPosOnBorder(const Vec3& vec3) const {
     if (vec3.y < -64 || vec3.y > 320) {
         return false;
     }
-
-    // 检查点是否在任何边上
-    for (size_t i = 0; i < mVertexs.size() - 1; ++i) {
-        const Vec3& v1 = mVertexs[i];
-        const Vec3& v2 = mVertexs[i + 1];
-
-        // 检查点是否在当前边上
-        if ((vec3.x >= std::min(v1.x, v2.x) && vec3.x <= std::max(v1.x, v2.x))
-            && (vec3.z >= std::min(v1.z, v2.z) && vec3.z <= std::max(v1.z, v2.z))) {
-            // 如果边是垂直的
-            if (v1.x == v2.x) {
-                if (vec3.x == v1.x) return true;
-            }
-            // 如果边是水平的
-            else if (v1.z == v2.z) {
-                if (vec3.z == v1.z) return true;
-            }
-            // 如果边是斜的（虽然在这个情况下不太可能）
-            // else {
-            //     double slope     = (v2.z - v1.z) / (v2.x - v1.x);
-            //     double intercept = v1.z - slope * v1.x;
-            //     if (std::abs(vec3.z - (slope * vec3.x + intercept)) < 1e-6) return true;
-            // }
-        }
-    }
-
-    return false;
+    return Polygon(mVertexs).isOnEdge(vec3);
 }
 bool PlotPos::isAABBOnBorder(BlockPos const& min, BlockPos const& max) const {
     if (!isValid()) {
         return false;
     }
-
-    // 检查Cube是否完全在地皮外部或内部
-    bool allInside  = true;
-    bool allOutside = true;
-    for (const auto& corner : PlotPos::getAABBVertexs(min, max)) {
-        bool inside  = isPosInPlot(corner);
-        allInside   &= inside;
-        allOutside  &= !inside;
-        if (!allInside && !allOutside) {
-            break;
-        }
-    }
-    if (allInside || allOutside) {
-        return false;
-    }
-
-    // 检查Cube的边是否与地皮边界相交
-    for (size_t i = 0; i < mVertexs.size() - 1; ++i) {
-        const BlockPos& v1 = mVertexs[i];
-        const BlockPos& v2 = mVertexs[i + 1];
-
-        // 检查水平边
-        if (v1.z == v2.z) {
-            if (min.z <= v1.z && max.z >= v1.z
-                && std::max(min.x, std::min(v1.x, v2.x)) <= std::min(max.x, std::max(v1.x, v2.x))) {
-                return true;
-            }
-        }
-        // 检查垂直边
-        else if (v1.x == v2.x) {
-            if (min.x <= v1.x && max.x >= v1.x
-                && std::max(min.z, std::min(v1.z, v2.z)) <= std::min(max.z, std::max(v1.z, v2.z))) {
-                return true;
-            }
-        }
-    }
-
-    return false;
+    return Polygon(mVertexs).isAABBOnEdge(min, max);
 }
 bool PlotPos::isRadiusOnBorder(BlockPos const& center, int radius) const {
     if (!isValid()) {
         return false;
     }
-
-    // 快速检查：如果圆心到地皮中心的距离大于半径加上地皮对角线的一半，则一定不相交
-    Vec3   plotCenter = (mVertexs[0] + mVertexs[2]) * 0.5;
-    double dx         = center.x - plotCenter.x;
-    double dz         = center.z - plotCenter.z;
-    double centerDist = std::sqrt(dx * dx + dz * dz);
-    double plotRadius = (mVertexs[2] - mVertexs[0]).length() * 0.5;
-    if (centerDist > radius + plotRadius) {
-        return false;
-    }
-
-    // 检查圆是否完全包含地皮或完全在地皮外部
-    bool allInside  = true;
-    bool allOutside = true;
-    for (const auto& vertex : mVertexs) {
-        dx                 = vertex.x - center.x;
-        dz                 = vertex.z - center.z;
-        double distSquared = dx * dx + dz * dz;
-        if (distSquared <= radius * radius) {
-            allOutside = false;
-        } else {
-            allInside = false;
-        }
-        if (!allInside && !allOutside) {
-            break;
-        }
-    }
-    if (allInside || allOutside) {
-        return false;
-    }
-
-    // 检查圆是否与地皮的边相交
-    for (size_t i = 0; i < mVertexs.size() - 1; ++i) {
-        const Vec3& v1 = mVertexs[i];
-        const Vec3& v2 = mVertexs[i + 1];
-
-        // 计算边的方向向量
-        double edgeX = v2.x - v1.x;
-        double edgeZ = v2.z - v1.z;
-
-        // 计算从圆心到边起点的向量
-        double vecX = center.x - v1.x;
-        double vecZ = center.z - v1.z;
-
-        // 计算边的长度的平方
-        double edgeLengthSquared = edgeX * edgeX + edgeZ * edgeZ;
-
-        // 计算圆心到边的投影长度比例
-        double t = (vecX * edgeX + vecZ * edgeZ) / edgeLengthSquared;
-        t        = std::max(0.0, std::min(1.0, t));
-
-        // 计算圆心到边的最近点
-        double nearestX = v1.x + t * edgeX;
-        double nearestZ = v1.z + t * edgeZ;
-
-        // 计算圆心到最近点的距离
-        double distX       = center.x - nearestX;
-        double distZ       = center.z - nearestZ;
-        double distSquared = distX * distX + distZ * distZ;
-
-        // 如果距离小于等于半径，则相交
-        if (distSquared <= radius * radius) {
-            return true;
-        }
-    }
-
-    return false;
+    return Polygon(mVertexs).isCircleOnEdge(center, radius);
 }
 string PlotPos::toString() const {
 #if !defined(DEBUG)
@@ -481,58 +354,7 @@ std::optional<PlotPos> PlotPos::tryMerge(PlotPos const& other) {
     if (!PlotPos::isAdjacent(*this, other)) {
         return std::nullopt;
     }
-
-    // 合并两个多边形的顶点
-    std::vector<Vec3> mergedVertexs = mVertexs;
-    mergedVertexs.insert(mergedVertexs.end(), other.mVertexs.begin(), other.mVertexs.end());
-
-    // 使用凸包算法获取合并后的顶点
-    std::vector<Vec3> sortedVertexs = mergedVertexs;
-
-    // 使用 Graham 扫描算法计算凸包
-    std::sort(sortedVertexs.begin(), sortedVertexs.end(), [](const Vec3& a, const Vec3& b) -> bool {
-        if (a.z != b.z) return a.z < b.z;
-        return a.x < b.x;
-    });
-
-    Vec3 pivot = sortedVertexs[0];
-    std::sort(sortedVertexs.begin() + 1, sortedVertexs.end(), [&](const Vec3& a, const Vec3& b) -> bool {
-        double angleA = std::atan2(a.z - pivot.z, a.x - pivot.x);
-        double angleB = std::atan2(b.z - pivot.z, b.x - pivot.x);
-        if (angleA == angleB) return (pivot - a).length() < (pivot - b).length();
-        return angleA < angleB;
-    });
-
-    std::vector<Vec3> hull;
-    hull.push_back(sortedVertexs[0]);
-    hull.push_back(sortedVertexs[1]);
-
-    for (size_t i = 2; i < sortedVertexs.size(); ++i) {
-        while (hull.size() >= 2) {
-            Vec3   q     = hull[hull.size() - 2];
-            Vec3   r     = hull[hull.size() - 1];
-            Vec3   s     = sortedVertexs[i];
-            double cross = (r.x - q.x) * (s.z - q.z) - (r.z - q.z) * (s.x - q.x);
-            if (cross > 0) break;
-            hull.pop_back();
-        }
-        hull.push_back(sortedVertexs[i]);
-    }
-
-    // 确保多边形闭合
-    if (hull.front() != hull.back()) {
-        hull.push_back(hull.front());
-    }
-
-    // 检查合并后的多边形是否有效
-    if (hull.size() < 4) { // 至少三个顶点加闭合点
-        return std::nullopt;
-    }
-
-    PlotPos mergedPlot;
-    mergedPlot.mVertexs = hull;
-
-    return mergedPlot;
+    return Polygon(mVertexs).tryMerge(Polygon(other.mVertexs));
 }
 
 bool PlotPos::operator!=(PlotPos const& other) const { return !(*this == other); }
@@ -550,30 +372,14 @@ bool PlotPos::isAdjacent(const PlotPos& plot1, const PlotPos& plot2) {
     return ((dx == 0 && dz == 1) || (dx == 1 && dz == 0)) && (plot1.isValid() && plot2.isValid());
 }
 bool PlotPos::isPointInPolygon(const Vec3& point, Vertexs const& polygon) {
-    // 判断点是否在多边形内部（射线法）
-    bool inside = false;
-    int  n      = (int)polygon.size();
-    for (int i = 0, j = n - 1; i < n; j = i++) {
-        if (((polygon[i].z <= point.z && point.z < polygon[j].z) || (polygon[j].z <= point.z && point.z < polygon[i].z))
-            && (point.x < (polygon[j].x - polygon[i].x) * (point.z - polygon[i].z) / (polygon[j].z - polygon[i].z)
-                              + polygon[i].x)) {
-            inside = !inside;
-        }
-    }
-    return inside;
+    return Polygon(polygon).isPointInPolygon(point);
 }
 int PlotPos::getSurfaceYStatic() {
     return TemplateManager::isUseTemplate() ? (TemplateManager::mTemplateData.template_offset + 1)
                                             : -64 + (Config::cfg.generator.subChunkNum * 16);
 }
 Vertexs PlotPos::getAABBVertexs(BlockPos const& min, BlockPos const& max) {
-    return {
-        Vec3{min.x, 0, min.z}, // 左下
-        Vec3{max.x, 0, min.z}, // 右下
-        Vec3{max.x, 0, max.z}, // 右上
-        Vec3{min.x, 0, max.z}, // 左上
-        Vec3{min.x, 0, min.z}  // 回到起点，形成闭合多边形
-    };
+    return Polygon::getAABBAroundVertexs(min, max);
 }
 std::vector<PlotPos> PlotPos::getPlotPosAt(BlockPos const& min, BlockPos const& max) {
     std::vector<PlotPos> rangedPlots;
@@ -655,25 +461,14 @@ std::vector<PlotPos> PlotPos::getPlotPosAt(BlockPos const& center, int radius) {
     return rangedPlots;
 }
 bool PlotPos::isAABBCollision(BlockPos const& min1, BlockPos const& max1, BlockPos const& min2, BlockPos const& max2) {
-    return !(
-        max1.x < min2.x || min1.x > max2.x || max1.y < min2.y || min1.y > max2.y || max1.z < min2.z || min1.z > max2.z
-    );
+    return Polygon::isAABBCollision(min1, max1, min2, max2);
 }
 void PlotPos::fillAABB(const BlockPos& min, const BlockPos& max, const Block& block) {
     auto& bs = ll::service::getLevel()->getDimension(getPlotWorldDimensionId())->getBlockSourceFromMainChunkSource();
 
     BlockPos start = min;
     BlockPos end   = max;
-
-    if (start.x > end.x) {
-        std::swap(start.x, end.x);
-    }
-    if (start.y > end.y) {
-        std::swap(start.y, end.y);
-    }
-    if (start.z > end.z) {
-        std::swap(start.z, end.z);
-    }
+    Polygon::fixAABB(start, end);
 
     for (int x = start.x; x <= end.x; ++x) {
         for (int z = start.z; z <= end.z; ++z) {
@@ -761,15 +556,12 @@ string  PlotCross::toString() const {
     return fmt::format("{} | {} => {}", getCrossID(), mDiagonPos.first.toString(), mDiagonPos.second.toString());
 }
 bool PlotCross::hasPoint(BlockPos const& pos) const {
-    int const&  x   = pos.x;
-    int const&  z   = pos.z;
-    auto const& min = mDiagonPos.first;
-    auto const& max = mDiagonPos.second;
-    return x >= min.x && x <= max.x && z >= min.z && z <= max.z;
+    return Polygon::isPointInAABB(pos, mDiagonPos.first, mDiagonPos.second);
 }
 void PlotCross::fill(Block const& block, bool removeBorder) {
     auto min = removeBorder ? mDiagonPos.first - 1 : mDiagonPos.first;
     auto max = removeBorder ? mDiagonPos.second + 1 : mDiagonPos.second;
+    Polygon::fixAABB(min, max);
 
     auto dim = ll::service::getLevel()->getDimension(getPlotWorldDimensionId());
     if (!dim) {
@@ -784,7 +576,7 @@ void PlotCross::fill(Block const& block, bool removeBorder) {
         for (int z = min.z; z <= max.z; z++) {
             auto& bl = bs.getBlock(x, y, z);
 
-            if (bl.isAir()) {
+            if (!bl.isAir()) {
                 bs.setBlock(x, y, z, block, (int)BlockUpdateFlag::AllPriority, nullptr);
             }
 
@@ -923,14 +715,14 @@ string PlotRoad::toString() const {
     return fmt::format("{} | {} => {}", getRoadID(), mDiagonPos.first.toString(), mDiagonPos.second.toString());
 }
 bool PlotRoad::hasPoint(BlockPos const& pos) const {
-    return pos.x >= mDiagonPos.first.x && pos.x <= mDiagonPos.second.x && pos.z >= mDiagonPos.first.z
-        && pos.z <= mDiagonPos.second.z;
+    return Polygon::isPointInAABB(pos, mDiagonPos.first, mDiagonPos.second);
 }
 void PlotRoad::fill(Block const& block, bool removeBorder) {
     bool const isX = mDirection == PlotDirection::East;
 
-    auto const& min = mDiagonPos.first;
-    auto const& max = mDiagonPos.second;
+    auto& min = mDiagonPos.first;
+    auto& max = mDiagonPos.second;
+    Polygon::fixAABB(min, max);
 
     auto dim = ll::service::getLevel()->getDimension(getPlotWorldDimensionId());
     if (!dim) {
